Allow the use of a different python interpreter
The project is python 2 based. On some distro (e.g. archlinux) the
default python interpreter is version 3, so the compilation fails with
the following error:

    python ../../objdictgen/objdictgen.py CANOpenShellMasterOD.od CANOpenShellMasterOD.c
      File "../../objdictgen/objdictgen.py", line 32
        print _("\nUsage of objdictgen.py :")
              ^
    SyntaxError: invalid syntax

Added an option to configure (that falls back to "python", so backward
compatibility is retained) that allows to override the command to use,
e.g. "python2" on archlinux.